payless-tax/withPersona/pyproject.toml
[project]
name = "payless-tax-demo"
version = "0.1.0"
description = "Payless Tax KYC Demo with Persona"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "streamlit>=1.30.0",
    "requests>=2.31.0",
    "mysql-connector-python>=8.2.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0"
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0"
]
